server/pyproject.toml
[project]
name = "app"
version = "1.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.11, <3.12"

dependencies = [
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "bcrypt (>=5.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "tenacity (>=9.1.2,<10.0.0)",
    "graphviz (>=0.21,<0.22)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.3,<4.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-asyncio (>=1.0.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pillow (>=10.0.0,<11.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.12"
black = "^25.1.0"
alembic = "^1.16.2"
aiosqlite = "^0.21.0"
pytest = "^8.4.2"
pytest-asyncio = "^1.2.0"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

